Warning correction. Level 4 with MSVC. Keeping option original content when updating option value.

+++ src/bitboard.cpp
@@ -111,7 +111,7 @@ void Bitboards::init() {
       for (Square s2 = SQ_A1; s2 <= SQ_H8; ++s2)
           if (s1 != s2)
           {
-              SquareDistance[s1][s2] = std::max(distance<File>(s1, s2), distance<Rank>(s1, s2));
+              SquareDistance[s1][s2] = static_cast<int8_t>( std::max(distance<File>(s1, s2), distance<Rank>(s1, s2)) );
               DistanceRingBB[s1][SquareDistance[s1][s2]] |= s2;
           }
